+++ src/test/java/AccountTest.java
@@ -18,7 +18,7 @@ public class AccountTest {
 
   @Test
   public void checkBalance() {
-    assertThat(account.getBalance(), is(1000));
+    assertThat(account.getBalance(), is((float) 1000));
   }
 
   @Test
